refactor slideshow -> login; maquetacion para el login
