+++ .idea/misc.xml
@@ -6,7 +6,11 @@
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/activity_landing_menu.xml" value="0.4413757165190203" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/activity_main.xml" value="0.33" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/adapter_menu_list.xml" value="0.24947916666666667" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_gallery.xml" value="0.31354166666666666" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_home.xml" value="0.6084860173577628" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_home_menu.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_login.xml" value="0.31354166666666666" />
+        <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/fragment_slideshow.xml" value="0.31354166666666666" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/layout/nav_header_landing_menu.xml" value="0.31354166666666666" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/menu/activity_landing_menu_drawer.xml" value="0.31354166666666666" />
         <entry key="..\:/Users/willi/AndroidStudioProjects/pizzeria_forkify/app/src/main/res/menu/landing_menu.xml" value="0.33" />
