+++ app/src/main/java/sv/edu/udb/pizzeria_forkify/OrderPizza/LandingMenuActivity.java
@@ -42,7 +42,7 @@ public class LandingMenuActivity extends AppCompatActivity {
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         mAppBarConfiguration = new AppBarConfiguration.Builder(
-                R.id.nav_menu, R.id.nav_gallery, R.id.nav_slideshow)
+                R.id.nav_menu, R.id.nav_gallery, R.id.nav_login)
                 .setOpenableLayout(drawer)
                 .build();
         NavController navController = Navigation.findNavController(this, R.id.nav_host_fragment_content_landing_menu);
