+++ app/src/main/java/sv/edu/udb/pizzeria_forkify/OrderPizza/ui/login/LoginViewModel.java
@@ -1,14 +1,14 @@
-package sv.edu.udb.pizzeria_forkify.OrderPizza.ui.slideshow;
+package sv.edu.udb.pizzeria_forkify.OrderPizza.ui.login;
 
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import androidx.lifecycle.ViewModel;
 
-public class SlideshowViewModel extends ViewModel {
+public class LoginViewModel extends ViewModel {
 
     private final MutableLiveData<String> mText;
 
-    public SlideshowViewModel() {
+    public LoginViewModel() {
         mText = new MutableLiveData<>();
         mText.setValue("This is slideshow fragment");
     }
