+++ app/src/main/res/menu/activity_landing_menu_drawer.xml
@@ -13,8 +13,8 @@
             android:icon="@drawable/ic_menu_gallery"
             android:title="@string/menu_gallery" />
         <item
-            android:id="@+id/nav_slideshow"
-            android:icon="@drawable/ic_menu_slideshow"
-            android:title="@string/menu_slideshow" />
+            android:id="@+id/nav_login"
+            android:icon="@drawable/user"
+            android:title="@string/menu_login" />
     </group>
 </menu>
