+++ app/src/main/res/values/strings.xml
@@ -10,5 +10,5 @@
 
     <string name="menu_home_pizzas_menu">Menu</string>
     <string name="menu_gallery">Gallery</string>
-    <string name="menu_slideshow">Slideshow</string>
+    <string name="menu_login">Inciar Sesion</string>
 </resources>
